+++ main.py
@@ -30,9 +30,6 @@ print('当前路径为'+current_path)
 
 # os.chdir('D:/1share/1item/Python/text_option/')
 
-
-
-
 def welcome():
     welcome_doc = open(current_path+'/doc/welcome.txt','r',encoding="utf-8")
     custome_doc = open(current_path+'/doc/custome.txt','r',encoding="utf-8")
@@ -69,99 +66,10 @@ def help():
         print(i)
     help_doc.close()
 
-#判断剪贴板的内容是否和上一次不相同，是则返回TRUE，否则返回FALSE
-def copy_notsame(input_data,old_input):
-    if input_data != old_input:
-        return True
-    else:
-        return False
-
-def count_words(Clipboard_text):
-    count = 0
-    for i in range(0,len(Clipboard_text)):
-        if Clipboard_text[i] in string.ascii_letters:
-            if i == (len(Clipboard_text)-1):
-                count = count +1
-            elif not Clipboard_text[i+1] in string.ascii_letters:
-                count = count +1
-            else:
-                continue
-        elif Clipboard_text[i].isalpha():
-            count = count +1
-        else:
-            continue
-    print("剪切板字数为{}".format(count))
-
-def count_characters(Clipboard_text):
-    count = len(Clipboard_text)
-    print("剪切板字符数为{}".format(count))
-
-#将内容中的大写替换为小写
-def Up2Low(Clipboard_text):
-    print('\n')
-    print(Clipboard_text.lower())
-    print('\n')
-    return Clipboard_text.lower()
-
-#将内容中的小写替换为大写
-def Low2Up(Clipboard_text):
-    print('\n')
-    print(Clipboard_text.upper())
-    print('\n')
-    return Clipboard_text.upper()
-
-#打开exe文件,路径需要先经过/处理
-#input = 'F:/1.software/Tencent/QQ/Bin/QQScLauncher.exe'
-def OpenExe(input):
-    for i in input[1:]:
-        try:
-            print('正在打开软件')
-            i = str(change(i))
-            for b in range(1,len(i)):
-                a = eval(str(-b))
-                if i[a] == '/':
-                    locate = a+1
-                    break
-                else:
-                    continue
-            software = i[locate:-4]
-            conmand = 'start '+ '"{1}" "{0}"'.format(i,software)
-            print(conmand)
-            os.system('chcp 65001')
-            os.system('@echo off')
-            os.system(conmand)
-        except:
-            print('打开软件错误')
-
-# 定义一个执行cmd脚本的命令
-def cmd(input):
-    for i in input[1:]:
-        try:
-            print('正在执行命令'+ i)
-            os.system('chcp 65001')
-            os.system(i)
-        except:
-            print('执行出现错误')
-
 # 定义一个退出命令
 def quite():
     exit()
  
-
-# 读取py文件夹的python脚本并执行
-def script(conmand):
-    for i in conmand[1:]:
-        file_path = current_path + '/py/'+ i 
-        python_path = current_path + '/python_env/python.exe'
-        conm = python_path +' ' + file_path
-        try:
-            print('正在打开'+ file_path)
-            #exec(open(file_path,encoding='utf-8').read())
-            os.system('chcp 65001')
-            os.system(conm)
-        except:
-            print('执行文件{}\n 出现错误'.format(file_path))
-
 # 定义复合指令解析
 def combind_conmand(input):
     for i in input[1:]:
@@ -179,7 +87,7 @@ def conmand_analys(input_conmand):
     m = methonds[c0]
     if m in ('Clipboard'):
         input = pyperclip.paste()
-        out = eval(c0+'(input)') #调用剪切板方法
+        out = eval('text_option.text_option.'+c0+'(input)') #调用剪切板方法，引入text_option.py模块
         if out == None:
             out = out
         else:
